pyG/requirements_embeddings.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=h0e60522_0
aiohttp=3.8.1=py310he2412df_1
aiosignal=1.3.1=pyhd8ed1ab_0
arrow-cpp=11.0.0=h2c9b28c_2
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
aws-c-common=0.4.57=ha925a31_1
aws-c-event-stream=0.1.6=hd77b12b_5
aws-checksums=0.1.9=ha925a31_0
aws-sdk-cpp=1.8.185=hd77b12b_0
blas=1.0=mkl
boost-cpp=1.78.0=h5b4e17d_0
bottleneck=1.3.5=py310h9128911_0
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2023.7.22=h56e8100_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=py310haa95532_0
cryptography=41.0.3=py310h3438e0d_0
cuda-cccl=12.2.140=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.14.5=pyhd8ed1ab_0
dill=0.3.7=pyhd8ed1ab_0
fftw=3.3.9=h2bbff1b_1
filelock=3.9.0=py310haa95532_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=py310h2bbff1b_0
fsspec=2023.6.0=pyh1a96a4e_0
geos=3.10.2=h39d44d4_0
gflags=2.2.2=ha925a31_1004
giflib=5.2.1=h8cc25b3_3
glog=0.5.0=h4797de2_0
grpc-cpp=1.48.2=hf108199_0
huggingface_hub=0.17.2=pyhd8ed1ab_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=py310haa95532_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.1.0=h57928b3_46319
jinja2=3.1.2=py310haa95532_0
joblib=1.2.0=py310haa95532_0
jpeg=9e=h2bbff1b_1
lerc=3.0=hd77b12b_0
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=12_win64_mkl
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.141=0
libcurand-dev=10.3.3.141=0
libcurl=8.2.1=h86230a5_0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
liblapack=3.9.0=12_win64_mkl
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libssh2=1.10.0=hcd4344a_2
libthrift=0.15.0=h636ae23_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py310h2bbff1b_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py310h2bbff1b_0
mpmath=1.3.0=py310haa95532_0
multidict=6.0.2=py310h2bbff1b_0
multiprocess=0.70.13=py310he2412df_1
neo4j-python-driver=5.8.0=pyhd8ed1ab_0
neobolt=1.7.17=py310he2412df_4
neomodel=4.0.8=pyhd8ed1ab_0
neotime=1.7.4=py_0
networkx=3.0=pypi_0
nltk=3.8.1=pyhd8ed1ab_0
numexpr=2.8.4=py310hd213c9f_0
numpy=1.22.3=py310hed7ac4c_2
openssl=1.1.1w=h2bbff1b_0
orc=1.7.4=h623e30f_1
packaging=23.1=py310haa95532_0
pandas=2.0.3=py310h4ed8f06_0
pillow=9.4.0=py310hd77b12b_1
pip=22.3.1=py310haa95532_0
psutil=5.9.0=py310h2bbff1b_0
pyarrow=11.0.0=py310h790e06d_1
pycparser=2.21=pyhd3eb1b0_0
pyg=2.3.1=py310_torch_2.0.0_cu117
pyopenssl=23.2.0=py310haa95532_0
pyparsing=3.0.9=py310haa95532_0
pysocks=1.7.1=py310haa95532_0
python=3.10.9=h966fe2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py310h2bbff1b_1
python_abi=3.10=2_cp310
pytorch=2.0.1=py3.10_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py310he2412df_4
re2=2022.04.01=h0e60522_0
regex=2022.7.9=py310h2bbff1b_0
requests=2.31.0=py310haa95532_0
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.3.0=py310h4ed8f06_0
scipy=1.7.3=py310h6d2d95c_2
sentence-transformers=2.2.2=pyhd8ed1ab_0
sentencepiece=0.1.99=py310h59b6b97_0
setuptools=65.6.3=py310haa95532_0
shapely=1.8.1=py310h3578588_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h6c2663c_0
sqlite=3.40.1=h2bbff1b_0
sympy=1.11.1=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.13.2=py310h49fca51_1
torchaudio=2.0.2=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.65.0=py310h9909e9c_0
transformers=4.28.1=pyhd8ed1ab_0
typing-extensions=4.7.1=py310haa95532_0
typing_extensions=4.7.1=py310haa95532_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.16=py310haa95532_0
utf8proc=2.6.1=h2bbff1b_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xxhash=0.8.0=h2bbff1b_3
xz=5.2.10=h8cc25b3_1
yaml=0.2.5=h8ffe710_2
yarl=1.7.2=py310he2412df_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
